Allow x, y, and z values to be transformed using the 3D depth effect
