+++ src/main/java/sh/ball/parser/lua/LuaSampleSource.java
@@ -18,69 +18,21 @@ import static sh.ball.gui.Gui.logger;
 
 public class LuaSampleSource implements FrameSource<Vector2> {
 
-  private final Globals globals;
-  private final Map<LuaString, LuaValue> bindings = new HashMap<>();
-  private final Set<LuaString> programDefinedVariables = new HashSet<>();
-  private LuaFunction lastWorkingFunction;
-  private String lastWorkingTextScript;
-  private LuaFunction function;
-  private String textScript;
+  private final LuaExecutor executor;
   private boolean active = true;
-  private long step = 1;
 
   public LuaSampleSource(Globals globals) {
-    this.globals = globals;
+    this.executor = new LuaExecutor(globals);
   }
 
-  public void setFunction(String textScript, LuaFunction function) {
-    if (lastWorkingFunction == null) {
-      lastWorkingFunction = function;
-      lastWorkingTextScript = textScript;
-    }
-    this.function = function;
-    this.textScript = textScript;
+  public void setScript(String textScript) {
+    executor.setScript(textScript);
   }
 
   @Override
   public Vector2 next() {
-    try {
-      boolean updatedFile = false;
-
-      globals.setmetatable(new BindingsMetatable(bindings));
-      bindings.put(LuaValue.valueOf("step"), LuaValue.valueOf((double) step));
-
-      LuaValue result;
-      try {
-        result = (LuaValue) LuaSampleSource.eval(function, globals);
-        if (!lastWorkingTextScript.equals(textScript)) {
-          lastWorkingFunction = function;
-          lastWorkingTextScript = textScript;
-          updatedFile = true;
-        }
-      } catch (Exception e) {
-        logger.log(Level.INFO, e.getMessage(), e);
-        result = (LuaValue) LuaSampleSource.eval(lastWorkingFunction, globals);
-      }
-      step++;
-
-      if (updatedFile) {
-        // reset variables
-        step = 1;
-        List<LuaString> variablesToRemove = new ArrayList<>();
-        bindings.keySet().forEach(name -> {
-          if (!programDefinedVariables.contains(name)) {
-            variablesToRemove.add(name);
-          }
-        });
-        variablesToRemove.forEach(bindings::remove);
-      }
-
-      return new Vector2(result.get(1).checkdouble(), result.get(2).checkdouble());
-    } catch (Exception e) {
-      logger.log(Level.INFO, e.getMessage(), e);
-    }
-
-    return new Vector2();
+    LuaValue result = executor.execute();
+    return new Vector2(result.get(1).checkdouble(), result.get(2).checkdouble());
   }
 
   @Override
@@ -107,110 +59,10 @@ public class LuaSampleSource implements FrameSource<Vector2> {
   }
 
   public void setVariable(String variableName, Object value) {
-    programDefinedVariables.add(LuaValue.valueOf(variableName));
-    bindings.put(LuaValue.valueOf(variableName), toLua(value));
+    executor.setVariable(variableName, value);
   }
 
   public void resetStep() {
-    step = 1;
-  }
-
-  /*
-MIT License
-
-Copyright (c) 2007 LuaJ. All rights reserved.
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit p ersons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-
-------------------------------
-
-The below functions have all been modified by James Ball to optimise Lua parsing
-on machines with poorer performance.
-
-
-   */
-
-  static class BindingsMetatable extends LuaTable {
-
-    BindingsMetatable(Map<LuaString, LuaValue> bindings) {
-      this.rawset(LuaValue.INDEX, new TwoArgFunction() {
-        public LuaValue call(LuaValue table, LuaValue key) {
-          LuaValue value = bindings.get(key.checkstring());
-          if (value == null) {
-            return LuaValue.NIL;
-          } else {
-            return value;
-          }
-        }
-      });
-      this.rawset(LuaValue.NEWINDEX, new ThreeArgFunction() {
-        public LuaValue call(LuaValue table, LuaValue key, LuaValue value) {
-          if (value == null || value.type() == LuaValue.TNIL) {
-            bindings.remove(key.checkstring());
-          } else {
-            bindings.put(key.checkstring(), value);
-          }
-          return LuaValue.NONE;
-        }
-      });
-    }
-  }
-
-  static private LuaValue toLua(Object javaValue) {
-    return javaValue == null? LuaValue.NIL:
-      javaValue instanceof LuaValue? (LuaValue) javaValue:
-        CoerceJavaToLua.coerce(javaValue);
-  }
-
-  static private Object toJava(LuaValue luajValue) {
-    return switch (luajValue.type()) {
-      case LuaValue.TNIL -> null;
-      case LuaValue.TSTRING -> luajValue.tojstring();
-      case LuaValue.TUSERDATA -> luajValue.checkuserdata(Object.class);
-      case LuaValue.TNUMBER -> luajValue.isinttype() ?
-        (Object) luajValue.toint() :
-        (Object) luajValue.todouble();
-      default -> luajValue;
-    };
-  }
-
-  static private Object toJava(Varargs v) {
-    final int n = v.narg();
-    switch (n) {
-      case 0: return null;
-      case 1: return toJava(v.arg1());
-      default:
-        Object[] o = new Object[n];
-        for (int i=0; i<n; ++i)
-          o[i] = toJava(v.arg(i+1));
-        return o;
-    }
-  }
-
-  static private Object eval(LuaFunction f, Globals g) throws ScriptException {
-    try {
-      f = f.getClass().newInstance();
-    } catch (Exception e) {
-      logger.log(Level.SEVERE, e.getMessage(), e);
-      throw new ScriptException(e);
-    }
-    f.initupvalue1(g);
-    return toJava(f.invoke(LuaValue.NONE));
+    executor.resetStep();
   }
 }
